PaddingLabelを廃止し、リポジトリ詳細画面で使うラベルを実装
