--- iOSEngineerCodeCheck/DetailRepository/PaddingLabel.swift
@@ -1,26 +1,0 @@
-import Foundation
-import UIKit
-
-class PaddingLabel: UILabel {
-
-    var padding: UIEdgeInsets = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
-
-    override func drawText(in rect: CGRect) {
-        let newRect = rect.inset(by: padding)
-        super.drawText(in: newRect)
-    }
-
-    override var intrinsicContentSize: CGSize {
-        var contentSize = super.intrinsicContentSize
-        contentSize.height += padding.top + padding.bottom
-        contentSize.width += padding.left + padding.right
-        return contentSize
-    }
-
-    override func sizeThatFits(_ size: CGSize) -> CGSize {
-        var contentSize = super.sizeThatFits(size)
-        contentSize.width += padding.left + padding.right
-        contentSize.height += padding.top + padding.bottom
-        return contentSize
-    }
-}
